requirements.txt
# Core dependencies
aiohttp>=3.8
python-dotenv>=1.0
pydantic>=2.0
pydantic-settings>=0.1
webrtcvad>=2.0.10
soundfile>=0.12.1
numpy>=1.24
requests>=2.31
dotenv
livekit>=0.8.0
livekit-agents>=0.11.0
langchain>=0.1.0
langchain-google-genai>=0.4.0
anthropic>=0.25.0
langchain-anthropic>=0.1.0
pyttsx3>=2.90
gTTS>=2.3.2
google-cloud-texttospeech>=2.14.0
google-generativeai>=0.3.0